Updated resource merging to properly handle nested require statements
Previous to this commit the merge process was not recursive. In the
implementation process I moved the writing of the merged content to the
output file out into the Application class. This is a better design because
it decouples the merge output of the Resource class from the filesystem.
This makes testing easier and allows for things like piping the merged
content to standard out.