Fixed Resource so that includes from multiple sources only appear once
Previous to this commit a dependency required by two files was included
twice. To handle this properly I first need to walk the require statements
in order to determine which requirements will be satified, then walk through
the requirements backwards so that merging a requirement later in the
file will not disturb the position of a requirement earlier in the file.