Add a Result base class for process outputs
I want the output of the forthcoming YUI
Compressor to share functionality with the output
of the JSLintAnalyzer so I created the Result base
class.

Setting up inheritance prompted me to add methods
for adding messages, warnings, and errors to a
Result instance and marking a Result instance as
good or bad.