Refactor Resource merging to expose Chunk list
I want to give Analyzer and Minifier instances the
ability to inspect the individual file chunks that
are merged together. This will allow a Minifier to 
be run against a Resource that mixed minified and
non-minified requirements and JSLintAnalyzer to
be run against a resource that merges libraries
that do not pass a lint check.