Rename Environment to Paths
Paths is a more appropriate name for what the class is, a collection of
search paths and an output path.

I also want to reserve the concept of an 'environment' for the future.
I plan to incorporate token substitution/template processing into Blend
and I want to have different environments ('dev', 'test') with different
sets of tokens.