Fix JSLintAnalyzer to inspect all JSLint process output
Matt M and I uncovered a bug in the original Azavea minifier
where the text output from the JSLint process is at a different
index in the output tuple depending on the runner (NodeJS vs. CScript)

Because I lifted the subprocess code straight from the original
Azavea minifier I needed to fix this bug as well.