Fix handling of dashed file names
Previous to this commit I was incorrectly treating
any dash suffix as a version number when computing
the base name of a resource. As a result, the
base name of 'jquery-1.2.3.js' was correctly
computed as 'jquery' but the base name of
'some-library.js' was incorrectly computed as
'some'

I now test that the first digit of the last
dashed segment is numeric before chopping it off
and considering it the version number.