Modify Analyzer and Minifier to accept config dict
On a large project, the YUICompressor was failing
because Java was running out of heap space. I
needed the ability to pass in a custom command line
argument from the config file. The easiest way to
implement this was to make the Analyzers and
Minifiers take an options dictionary instead o
named properties.